requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.3.0
cffi==1.15.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
Django==4.0.5
django-debug-toolbar==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-spectacular==0.22.1
inflection==0.5.1
jsonschema==4.6.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
